Create alternative graph drawing block
This is a version I wrote before the current versions. It is simpler and works well for most graphs. If you want to play around with it, set the function to "x**3", the dimensions to 252, the bounds (-50,50) for domain and range and try different values for the connector variable. You can use the same setup to play with "scalar" variable in the main code block